# internal/api/openapi.yaml
openapi: "3.1.0"
info:
  title: sqi-server API
  description: |
    REST API for the sqi distributed task management server.

    All error responses use the RFC 7807 problem-details format
    (`application/problem+json`).  Clients should branch on `Content-Type`
    rather than status code alone to detect error bodies.
  version: "0.1.0"
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /api/v1
    description: Current server
...
components:
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource UUID
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 50
      description: Maximum number of items to return (1–1000).
    offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Zero-based item offset for pagination.
    sortDir:
      name: sort_dir
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
      description: Sort direction.
  schemas:
...
    ProblemDetail:
      type: object
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          example: "about:blank"
        title:
          type: string
          example: Not Found
        status:
          type: integer
          example: 404
        detail:
          type: string
          example: job not found
        instance:
          type: string
          description: Request-ID that identifies this specific occurrence.
          example: a1b2c3d4e5f60708

    # ── Jobs ─────────────────────────────────────────────────────────────────
...
    Task:
      type: object
      required:
        - id
        - job_id
        - step_id
        - name
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
        job_id:
          type: string
        step_id:
          type: string
        name:
          type: string
        parameters:
          type: object
          additionalProperties:
            type: string
          description: Resolved OpenJD parameter values for this task.
        status:
          type: string
          enum: [pending, ready, assigned, running, succeeded, failed, canceled]
        assigned_worker_id:
          type: string
        assigned_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TaskList:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    TaskLog:
      type: object
      required: [id, task_id, attempt_id, seq_num, nats_seq, stream, data, at, received_at]
      properties:
        id:
          type: string
        task_id:
          type: string
        attempt_id:
          type: string
        seq_num:
          type: integer
          format: int64
          description: Monotonic sequence number within the attempt.
        nats_seq:
          type: integer
          format: int64
          description: NATS JetStream sequence; use as `after_nats_seq` cursor.
        stream:
          type: string
          enum: [stdout, stderr]
        data:
          type: string
          description: Raw log text for this chunk.
        at:
          type: string
          format: date-time
          description: Timestamp assigned by the worker.
        received_at:
          type: string
          format: date-time
          description: Timestamp when the server persisted this chunk.
    TaskLogsResponse:
      type: object
      required: [items, after_nats_seq, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TaskLog"
        after_nats_seq:
          type: integer
          format: int64
          description: Echo of the `after_nats_seq` query parameter used for this page.
        limit:
          type: integer
    RetryResponse:
      type: object
      required: [task_id, status]
      properties:
        task_id:
          type: string
        status:
          type: string
          example: ready

    # ── Workers ───────────────────────────────────────────────────────────────
...
  responses:
...
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    Conflict:
      description: Conflict — unique constraint violated or invalid state transition.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
    InternalServerError:
      description: Internal server error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"

# ── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /jobs/{id}/tasks:
    parameters:
      - $ref: "#/components/parameters/id"

    get:
      operationId: listJobTasks
      summary: List tasks for a job
      tags: [tasks]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, ready, assigned, running, succeeded, failed, canceled]
          description: Filter by task status.
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [created_at, status, updated_at, name]
            default: created_at
        - $ref: "#/components/parameters/sortDir"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: Paginated list of tasks.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskList"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ── Tasks ──────────────────────────────────────────────────────────────────
  /tasks/{id}:
    parameters:
      - $ref: "#/components/parameters/id"

    get:
      operationId: getTask
      summary: Get a task
      tags: [tasks]
      responses:
        "200":
          description: Task detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /tasks/{id}/logs:
    parameters:
      - $ref: "#/components/parameters/id"

    get:
      operationId: getTaskLogs
      summary: Get task logs
      tags: [tasks]
      description: |
        Returns log chunks for the latest attempt of the given task.

        **Offset-based pagination**: pass the `nats_seq` of the last received
        item as `after_nats_seq` on successive calls to walk forward without
        repeating chunks.

        **Tail streaming**: when `tail=true` the response uses chunked transfer
        encoding (`Content-Type: application/x-ndjson`).  Each flushed line is a
        newline-delimited JSON `TaskLog` object.  Streaming stops when the task
        reaches a terminal state and all buffered chunks have been delivered, or
        when the client disconnects.

        When no attempt exists yet an empty result is returned (rather than 404)
        so pollers do not need to handle that case specially.
      parameters:
        - name: after_nats_seq
          in: query
          schema:
            type: integer
            format: int64
            default: 0
          description: Return chunks with `nats_seq` greater than this value.
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
          description: Maximum chunks per page (1–1000).
        - name: tail
          in: query
          schema:
            type: boolean
            default: false
          description: When true, stream chunks using newline-delimited JSON.
      responses:
        "200":
          description: |
            Log chunks.  When `tail=false` the body is a `TaskLogsResponse`
            JSON object.  When `tail=true` the body is a stream of
            newline-delimited `TaskLog` JSON objects
            (`Content-Type: application/x-ndjson`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskLogsResponse"
            application/x-ndjson:
              schema:
                $ref: "#/components/schemas/TaskLog"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /tasks/{id}/retry:
    parameters:
      - $ref: "#/components/parameters/id"

    post:
      operationId: retryTask
      summary: Retry a task
      tags: [tasks]
      description: |
        Resets a failed or canceled task to `ready` so the scheduler can
        pick it up on the next dispatch cycle.

        Only tasks in `failed` or `canceled` state may be retried; all other
        states are rejected with 409 Conflict.
      responses:
        "202":
          description: Task queued for retry.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetryResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ── Workers ────────────────────────────────────────────────────────────────
...